feat: Add Google Services configuration and update string resources
This commit introduces the `google-services.template.json` file, which contains project information and client details for Google services integration. It also includes an update to the `strings.xml` file, adding a new string resource `default_web_client_id`.

Additionally, minor type casting adjustments were made in `GoogleApiService.kt` and `EvidenceRepositoryImpl.kt` related to evidence handling.

Key Changes:

*   **New File**:
    *   `app/google-services.template.json`: Added template file for Google Services configuration, including project ID, storage bucket, and client information.
*   **String Resources (`app/src/main/res/values/strings.xml`)**:
    *   Added `default_web_client_id` string resource.
*   **`GoogleApiService.kt`**:
    *   In `addEvidenceToCase`, explicitly cast `values` to `List<List<Any>>` before calling `appendData`.
*   **`EvidenceRepositoryImpl.kt`**:
    *   In `deleteEvidence`, cast `evidence.id` to `Int` when calling `googleApiService?.deleteEvidenceFromCase`.

+++ app/src/main/java/com/hereliesaz/lexorcist/data/EvidenceRepositoryImpl.kt
@@ -40,7 +40,7 @@ class EvidenceRepositoryImpl(
     }
 
     override suspend fun deleteEvidence(spreadsheetId: String, evidence: Evidence) {
-        googleApiService?.deleteEvidenceFromCase(spreadsheetId, evidence.id)
+        googleApiService?.deleteEvidenceFromCase(spreadsheetId, evidence.id.toInt())
         refreshEvidenceForCase(spreadsheetId, evidence.caseId.toInt())
     }
 }
